requirements.txt
Django==4.2.0  # Фреймворк
djangorestframework==3.14.0  # Для создания REST API
psycopg2-binary==2.9.6  # Адаптер для PostgreSQL
Pillow==9.5.0  # Для работы с изображениями (загрузка мемов)
django-cors-headers==4.0.0  # Для обработки CORS
drf-spectacular==0.26.2  # Для автоматической документации API (Swagger)
djangorestframework-simplejwt==5.2.2  # Для JWT аутентификации в API
celery==5.2.7  # Для фоновых задач (рассылка) - опционально
redis==4.5.4  # Брокер для Celery/кэш - опционально
python-dotenv==1.0.0  # Для загрузки переменных окружения
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
